BUNDLE-751 Qa fixes (#77)

+++ database/UserAssetRepository.go
@@ -35,14 +35,14 @@ func (repo *UserAssetRepository) GetAssetsByID(id, model interface{}) error {
 }
 
 // GetAssetsByID ...
-func (repo *UserAssetRepository) SumAmountField(model interface{}) (int64, error) {
-	//var sum int
+func (repo *UserAssetRepository) SumAmountField(model interface{}) (float64, error) {
+	//var sum float64
 	//Note i am summing here using sql here so addition is in crypto decimal units which is what its saved in.
 	// This is fine for float management but dont use this method for transactional stuff. Floating point addition
 	// is a problem. rater convert to native units and then sum. :)
 
 	type NResult struct {
-		N int64 //or int ,or some else
+		N float64 //or int ,or some else
 	}
 
 	var n NResult
